+++ parametrization.py
@@ -15,15 +15,30 @@ SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
-from constants import inverter_name_dict
 from munch import Munch
-from viktor.parametrization import GeoPointField
-from viktor.parametrization import NumberField
-from viktor.parametrization import OptionField
-from viktor.parametrization import Parametrization
-from viktor.parametrization import Step
-from viktor.parametrization import Text
-from viktor.parametrization import ToggleButton
+from viktor.errors import UserError
+from viktor.geometry import GeoPoint
+from viktor.parametrization import (
+    GeoPointField,
+    NumberField,
+    OptionField,
+    Parametrization,
+    Step,
+    Text,
+    ToggleButton,
+)
+
+from constants import inverter_name_dict
+
+DEFAULT_LOCATION = GeoPoint(37.443889, -6.259444)
+
+
+def validate_step_1(params, **kwargs):
+    """Validates step 1."""
+    if not params.step_1.point:
+        raise UserError("Select a location on the map.")
+    if params.step_1.surface == 0:
+        raise UserError("The surface area should be larger than zero.")
 
 
 def _get_inverter_name_list(params: Munch, **kwargs):
@@ -39,43 +54,68 @@ def _get_inverter_name_list(params: Munch, **kwargs):
 class ConfiguratorParametrization(Parametrization):
     """Defines the input fields for the mapview (step 1) and dataview (step 2)"""
 
-    step_1 = Step("Step 1 Define location of your home", views="get_map_view")
+    step_1 = Step(
+        "Step 1 Define location of your home",
+        views=["get_map_view", "get_weather_data"],
+        on_next=validate_step_1,
+    )
 
     step_1.text = Text(
-        """## Location definition
-        \n Different locations around the globe interact with the sun differently.
-        \n Moreover the different climates produce specific weather patterns.
-        \n By choosing the location of your home on the map on the right, weather data for a
-        *Typical Meteorological Year* (TMY) at your location will be gathered from the
-        [EU Science Hub](https://ec.europa.eu/jrc/en/pvgis). This TMY is used in the calculation of energy yield by your
-        system."""
+        """# Welcome to the Solar configurator app!
+This app allows one to design and calculate the return-on-investment for a PV system in a simplified manner.
+
+## Location definition
+
+Different locations around the globe interact with the sun differently.
+
+Moreover the different climates produce specific weather patterns.
+
+By choosing the location of your home on the map on the right, weather data for a
+*Typical Meteorological Year* (TMY) at your location will be gathered from the
+[EU Science Hub](https://ec.europa.eu/jrc/en/pvgis). This TMY is used in the calculation of energy yield by your
+system.
+
+## Choose location
+"""
     )
-    step_1.text2 = Text("""## Choose location""")
-    step_1.point = GeoPointField("enter a point", description='Click a location on the map to select it for calculation')
+    step_1.point = GeoPointField(
+        "enter a point",
+        description="Click a location on the map to select it for calculation",
+        default=DEFAULT_LOCATION,
+    )
     step_1.text3 = Text("""Define the surface area available on the roof:""")
-    step_1.surface = NumberField("Surface area", suffix="m2", default=0, min=0,
-                                 description='Use the arrows to select your available surface area or enter a number '
-                                             '  \n (if applicable use a decimal point **\' . \'** instead of a comma **\' , \'** )')
+    step_1.surface = NumberField(
+        "Surface area",
+        suffix="m2",
+        default=20,
+        min=10,
+        max=200,
+        description="Use the arrows to select your available surface area or enter a number "
+        "  \n (if applicable use a decimal point **' . '** instead of a comma **' , '** )",
+    )
 
     step_2 = Step("Step 2 Choose your system configuration", views="get_data_view")
 
     step_2.text = Text(
         """## PV-System explanation
-        \n A consumer-home PV-system always consist of
-        \n - One inverter unit
-        \n - One or more modules
-        \n The PV-systems included here are all approved by either the
-        [Sandia National Laboratory](https://tinyurl.com/4jf5nkpy) or the
-        [California Energy Commission](https://tinyurl.com/2p87uwkj), both of which use different calculations
-        for their approval protocols."""
+A consumer-home PV-system always consist of
+- One inverter unit
+- One or more modules
+The PV-systems included here are all approved by either the
+[Sandia National Laboratory](https://tinyurl.com/4jf5nkpy) or the
+[California Energy Commission](https://tinyurl.com/2p87uwkj), both of which use different calculations
+for their approval protocols.
+"""
     )
     step_2.text2 = Text(
         """## Choose PV-System configuration
-        \n By choosing either institution as the *System type* one is able to choose different
-        configurations of inverters and modules.
-        \n All up-to-date secifications of the chosen system configuration are
-        provided by the *System Advisor Model* (SAM) as developed by the
-        [National Renewable Energy Laboratory](https://sam.nrel.gov/)."""
+By choosing either institution as the *System type* one is able to choose different 
+configurations of inverters and modules.
+
+All up-to-date secifications of the chosen system configuration are
+provided by the *System Advisor Model* (SAM) as developed by the
+[National Renewable Energy Laboratory](https://sam.nrel.gov/).
+"""
     )
 
     step_2.system_type = OptionField(
@@ -89,7 +129,7 @@ class ConfiguratorParametrization(Parametrization):
     step_2.inverter_name = OptionField(
         "Inverter model",
         options=_get_inverter_name_list,
-        default="Generac Power Systems Inverter",
+        default="ABB: PVI-0.3 Inverter",
         flex=50,
         autoselect_single_option=True,
     )
@@ -115,10 +155,12 @@ class ConfiguratorParametrization(Parametrization):
     step_3 = Step("Step 3 Visualise your return-on-investment", views="get_plotly_view")
     step_3.text = Text(
         """## Forecast and Break-even
-        \n Here you are able to forecast the energy yield of your chosen system. Based on the **kWh price** indicated
-        The forecast will automatically calculate the revenue produced by your system.
-        \n The break-even point is also indicated, this is the moment in time where the investment will be fully
-        compensated by the revenue produced."""
+Here you are able to forecast the energy yield of your chosen system. Based on the **kWh price** indicated
+The forecast will automatically calculate the revenue produced by your system.
+
+The break-even point is also indicated, this is the moment in time where the investment will be fully
+compensated by the revenue produced.
+"""
     )
     step_3.forecast_horizon = NumberField(
         "Enter the forecasting horizon",
@@ -126,9 +168,7 @@ class ConfiguratorParametrization(Parametrization):
         default=5,
         flex=80,
         min=1,
-        description="Amount of years starting at"
-        "  \n the beginning of this year"
-        "  \n (enter only whole years)",
+        description="Amount of years starting at" "  \n the beginning of this year" "  \n (enter only whole years)",
     )
     step_3.kwh_cost = NumberField(
         "Enter kWh price to calculate break-even",
@@ -141,12 +181,16 @@ class ConfiguratorParametrization(Parametrization):
         "  \n Prices may differ between energy providers."
         " \n"
         "  \n Note that prices for kWh supplied **to** the grid often"
-        " differ from those of kWh's **taken from** the grid",
+        " differ from those of kWhs **taken from** the grid",
     )
     step_3.text2 = Text(
         """If you are only interested in a detailed time period you could choose to not show the break-even point
-        in the graph and choose for a short forecasting horizon.
-        \n Alternatively you could use the viewing-tools in the top-right corner of the graph to zoom to a specific
-        point."""
+in the graph and choose for a short forecasting horizon.
+
+Alternatively you could use the viewing-tools in the top-right corner of the graph to zoom to a specific
+point.
+"""
     )
     step_3.break_even_toggle = ToggleButton("Show break-even point", default=True)
+
+    final_step = Step("What's next?", views="final_step")