+++ CHANGELOG.md
@@ -1,6 +1,29 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased] - yyyy-mm-dd
+### Added
+- Weather data added.
+- Energy yield for configuration added.
+- Vertical line and break-even duration added.
+- Step validation added.
+
+### Changed
+- Changed from 'simple' to 'editor' app
+- Bumped SDK version to v13.7.0
+- Refactored some code/text
+- README updated.
+- Constrained surface area to keep design realistic.
+
+### Deprecated
+None.
+
+### Removed
+None.
+
+### Fixed
+- Default for inverter.
+
 ## [Unreleased] - 2022-02-07
 ### Added
 - Manifest